CSDAIv2/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
openai==1.3.5
httpx>=0.27.0
scikit-learn==1.3.2
joblib==1.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
